removed type cast

+++ conv-harness.c
@@ -345,7 +345,7 @@ void student_conv(float *** image, int16_t **** kernels, float *** output,
   for (int i = 0; i < total_width; i++) {
     for (int j = 0; j < total_height; j++) {
       for (int k = 0; k < nchannels; k++) {
-        image_buffer[calc_image_index(i, j, k, total_width, total_height, nchannels)] = (double)image[i][j][k];
+        image_buffer[calc_image_index(i, j, k, total_width, total_height, nchannels)] = image[i][j][k];
       }
     }
   }
@@ -354,7 +354,7 @@ void student_conv(float *** image, int16_t **** kernels, float *** output,
     for (int j = 0; j < nchannels; j++) {
       for (int k = 0; k < kernel_order; k++) {
         for (int l = 0; l < kernel_order; l++) {
-          kernels_buffer[calc_kernels_index(i, j, k, l, nkernels, nchannels, kernel_order)] = (double)kernels[i][j][k][l];
+          kernels_buffer[calc_kernels_index(i, j, k, l, nkernels, nchannels, kernel_order)] = kernels[i][j][k][l];
         }
       }
     }
